TOOLS-381 use signed integer for bsondump type with debug
Former-commit-id: 6aadb2e21f92b451d7125a0a918e2555a353b273

+++ bsondump/bsondump.go
@@ -121,7 +121,7 @@ func DebugBSON(raw bson.Raw, indentLevel int, out io.Writer) error {
 		// 3. The BSON value
 		// So size == 1 [size of type byte] +  1 [null byte for cstring key] + len(bson key) + len(bson value)
 		// see http://bsonspec.org/spec.html for more details
-		fmt.Fprintf(out, "%v\t\t\ttype: %4v size: %v\n", indent, rawElem.Value.Kind,
+		fmt.Fprintf(out, "%v\t\t\ttype: %4v size: %v\n", indent, int8(rawElem.Value.Kind),
 			2+len(rawElem.Name)+len(rawElem.Value.Data))
 
 		//For nested objects or arrays, recurse.
